[ty] Extract `relation` module from `types.rs`
